Apply frontend conventions to form-builder module
