+++ clients/admin-ui/src/features/properties/privacy-center-config/form-builder/fieldUtils.tsx
@@ -1,4 +1,9 @@
-import { useEffect, useRef } from "react";
+import React, { useEffect, useRef } from "react";
+
+export interface IdentityFieldProps {
+  required: boolean;
+  "data-element-id"?: string;
+}
 
 export interface BaseFieldProps {
   name: string;
@@ -18,6 +23,20 @@ export const isEmptyValue = (v: unknown) =>
 
 export const stableValueKey = (v: unknown) => JSON.stringify(v ?? null);
 
+export const FieldWrapper = ({
+  elementId,
+  children,
+}: {
+  elementId?: string;
+  children: React.ReactNode;
+}) => {
+  if (!elementId) {
+    // eslint-disable-next-line react/jsx-no-useless-fragment
+    return <>{children}</>;
+  }
+  return <span data-element-id={elementId}>{children}</span>;
+};
+
 /**
  * Seed the field's binding from `default_value`:
  * - On first mount when nothing is in state yet.
@@ -27,7 +46,7 @@ export const stableValueKey = (v: unknown) => JSON.stringify(v ?? null);
  *
  * If the end user has typed something different, we leave their value alone.
  */
-export const useDefaultValueSeed = <T>(
+export const useDefaultValueSeed = <T,>(
   value: T | undefined,
   setValue: (next: T) => void,
   defaultValue: T | undefined,
