+++ clients/admin-ui/src/features/properties/privacy-center-config/form-builder/registry.tsx
@@ -4,7 +4,12 @@ import dynamic from "next/dynamic";
 import React from "react";
 
 import { catalog } from "./catalog";
-import { BaseFieldProps, useDefaultValueSeed } from "./fieldUtils";
+import {
+  BaseFieldProps,
+  FieldWrapper,
+  IdentityFieldProps,
+  useDefaultValueSeed,
+} from "./fieldUtils";
 
 // LocationField pulls in fidesui's LocationSelect, which transitively imports
 // iso-3166 (CJS). Turbopack rejects that on the SSR path with "CJS module
@@ -25,20 +30,6 @@ const PhoneFieldDynamic = dynamic(
 const FormContainer = ({ children }: { children?: React.ReactNode }) => (
   <Form layout="vertical">{children}</Form>
 );
-
-const FieldWrapper = ({
-  elementId,
-  children,
-}: {
-  elementId?: string;
-  children: React.ReactNode;
-}) => {
-  if (!elementId) {
-    // eslint-disable-next-line react/jsx-no-useless-fragment
-    return <>{children}</>;
-  }
-  return <span data-element-id={elementId}>{children}</span>;
-};
 
 // Each field binds its current value to /form/<name> in the json-render
 // state model. In Preview mode this lets visibility conditions react to
@@ -115,11 +106,6 @@ const MultiSelectField = ({ props }: { props: BaseFieldProps }) => {
     </FieldWrapper>
   );
 };
-
-interface IdentityFieldProps {
-  required: boolean;
-  "data-element-id"?: string;
-}
 
 const EmailField = ({ props }: { props: IdentityFieldProps }) => {
   const [value, setValue] = useFieldBinding<string>("email");
