+++ clients/admin-ui/src/features/properties/privacy-center-config/form-builder/useFormBuilder.ts
@@ -3,107 +3,15 @@ import { useCallback, useEffect, useRef, useState } from "react";
 import { useAppSelector } from "~/app/hooks";
 import { selectToken } from "~/features/auth/auth.slice";
 
-import { catalog } from "./catalog";
 import { streamChatTurn } from "./streaming";
-import type { JsonRenderSpec } from "./types";
-
-const KNOWN_TYPES = new Set(Object.keys(catalog.components));
-
-/**
- * Defense-in-depth: strip elements whose `type` is not in the catalog before
- * rendering in the preview. The LLM could return arbitrary component types;
- * unknown types would be no-ops in json-render, but removing them explicitly
- * keeps the spec clean and avoids surprises.
- */
-const sanitizeSpec = (spec: JsonRenderSpec): JsonRenderSpec => {
-  const root = spec.elements[spec.root];
-  if (!root) {
-    return spec;
-  }
-  const validChildren = root.children.filter((id) => {
-    const el = spec.elements[id];
-    return el && KNOWN_TYPES.has(el.type);
-  });
-  const cleanElements: JsonRenderSpec["elements"] = {
-    [spec.root]: { ...root, children: validChildren },
-  };
-  validChildren.forEach((id) => {
-    cleanElements[id] = spec.elements[id];
-  });
-  return { ...spec, elements: cleanElements };
-};
-
-export type ChatRole = "user" | "assistant" | "system";
-
-export interface ChatMessage {
-  role: ChatRole;
-  content: string;
-}
-
-export type Status = "idle" | "streaming" | "aborted" | "error";
-
-interface UseFormBuilderInput {
-  propertyId: string;
-  actionPolicyKey: string;
-  initialSpec: JsonRenderSpec | null;
-}
-
-export interface UseFormBuilder {
-  spec: JsonRenderSpec | null;
-  messages: ChatMessage[];
-  status: Status;
-  error: string | null;
-  sendMessage: (text: string) => Promise<void>;
-  abort: () => void;
-  setSpec: (spec: JsonRenderSpec | null) => void;
-}
-
-// Strip ```json … ``` fences and return the first balanced {...} object
-// the LLM wrote. Falls back to the original string if no braces are found.
-const extractJson = (raw: string): string => {
-  let body = raw.trim();
-  const fence = body.match(/^```(?:json)?\s*\n?([\s\S]*?)\n?```$/);
-  if (fence) {
-    body = fence[1].trim();
-  }
-
-  const firstBrace = body.indexOf("{");
-  if (firstBrace === -1) {
-    return body;
-  }
-
-  let depth = 0;
-  let inString = false;
-  let escape = false;
-  for (let i = firstBrace; i < body.length; i += 1) {
-    const ch = body[i];
-    if (escape) {
-      escape = false;
-    } else if (ch === "\\") {
-      escape = true;
-    } else if (ch === '"') {
-      inString = !inString;
-    } else if (!inString) {
-      if (ch === "{") {
-        depth += 1;
-      } else if (ch === "}") {
-        depth -= 1;
-        if (depth === 0) {
-          return body.slice(firstBrace, i + 1);
-        }
-      }
-    }
-  }
-  return body.slice(firstBrace);
-};
-
-const tryParse = (raw: string): JsonRenderSpec | null => {
-  try {
-    return JSON.parse(extractJson(raw)) as JsonRenderSpec;
-  } catch {
-    return null;
-  }
-};
+import type {
+  ChatMessage,
+  JsonRenderSpec,
+  Status,
+  UseFormBuilder,
+  UseFormBuilderInput,
+} from "./types";
+import { sanitizeSpec, tryParseSpec } from "./utils";
 
 export function useFormBuilder(input: UseFormBuilderInput): UseFormBuilder {
   const [spec, setSpec] = useState<JsonRenderSpec | null>(input.initialSpec);
@@ -162,7 +70,7 @@ export function useFormBuilder(input: UseFormBuilderInput): UseFormBuilder {
         for await (const ev of stream) {
           if (ev.event === "chunk") {
             buffer += ev.data;
-            const parsed = tryParse(buffer);
+            const parsed = tryParseSpec(buffer);
             if (parsed) {
               const sanitized = sanitizeSpec(parsed);
               specRef.current = sanitized;
@@ -176,7 +84,7 @@ export function useFormBuilder(input: UseFormBuilderInput): UseFormBuilder {
               // structured server payload – if it isn't valid JSON, ignore it
             }
             if (payload?.raw) {
-              const final = tryParse(payload.raw);
+              const final = tryParseSpec(payload.raw);
               if (final) {
                 const sanitized = sanitizeSpec(final);
                 specRef.current = sanitized;
